backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
httpx==0.27.2
pdfplumber==0.10.3
pdf2image==1.16.3
Pillow==10.4.0
python-dotenv==1.0.0



